rename package to hcloudmock
