+++ hcloud/exp/hcloudmock/zz_action_client_mock.go
@@ -7,7 +7,7 @@
 //
 
 // Package mock is a generated GoMock package.
-package mock
+package hcloudmock
 
 import (
 	context "context"
